+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
 scheduling.mysql.enabled=true
-scheduling.mysql.url=jdbc:mysql://localhost:3306/testdb?useUnicode=true&characterEncoding=UTF-8&serverTimezone=Asia/Shanghai&useSSL=false
+scheduling.mysql.url=jdbc:mysql://127.0.0.1:3306/testdb?useUnicode=true&characterEncoding=UTF-8&serverTimezone=Asia/Shanghai&useSSL=false
 scheduling.mysql.username=root
-scheduling.mysql.password=admin
+scheduling.mysql.password=deipss
