+++ src/test/java/org/deipss/scheduling/db/impl/TaskB.java
@@ -1,17 +1,18 @@
-package org.deipss.scheduling.service.impl;
+package org.deipss.scheduling.db.impl;
 
 import lombok.extern.slf4j.Slf4j;
 import org.deipss.scheduling.service.AbstractTask;
 import org.deipss.scheduling.util.SleepUtil;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.stereotype.Component;
 
 @Slf4j
 @Component
-public class TaskA extends AbstractTask {
+public class TaskB extends AbstractTask {
     @Override
     public Boolean doBiz() {
-        SleepUtil.sleepSecond(3);
-        log.info("任务运行A");
-        return true;
+        SleepUtil.sleepSecond(2);
+        log.info("任务运行B");
+        return false;
     }
 }
